Changed all vectors to float3 instead of double3

+++ RenderThis/Point.swift
@@ -9,4 +9,4 @@
 import Foundation
 import GLKit
 
-typealias Point = double3
+typealias Point = float3
